+++ analyzefit/__init__.py
@@ -3,3 +3,6 @@ import os
 
 if os.system != "nt": #pragma: no cover
     matplotlib.use("Agg")
+
+
+from analyzefit.analyze import Analysis