+++ HISTORY.md
@@ -4,6 +4,9 @@
 - Added additional unit testing functions for travis.
 - Changed import of plotting subroutine to import for analyzefit explicitly.
 - Changed import of manipulate subroutine to import for analyzefit explicitly.
+- Fixed module name in tox.in.
+- Added pypi and quantifiedcode badges.
+- Fixed last issues found by quantifiedcode.
 
 ## Revision 0.3.3
 - Fixed bug reported in Issue #5, the number of y values and predictions must now agree.